+++ src/semantic/checks/block.rs
@@ -1,8 +1,18 @@
 use crate::{
     error_handling::MistiError,
     semantic::{impls::SemanticCheck, symbol_table::SymbolTable},
-    syntax::ast::BlockMember,
+    syntax::ast::{Block, BlockMember},
 };
+
+impl SemanticCheck for Block<'_> {
+    fn check_semantics(&self, scope: &SymbolTable) -> Result<(), MistiError> {
+        for member in &self.members {
+            member.check_semantics(scope)?;
+        }
+
+        Ok(())
+    }
+}
 
 impl<'a> SemanticCheck for BlockMember<'a> {
     // TODO: A block may contain a function declaration statement,