+++ src/semantic/checks/mod.rs
@@ -1,5 +1,7 @@
 pub mod binding;
 pub mod block;
+pub mod conditional;
 pub mod expression;
+pub mod for_loop;
 pub mod function_declaration;
 pub mod top_level_declaration;